NDRS-650: Move `FinalitySignature` type to a common crate.
