+++ node/src/effect/requests.rs
@@ -42,13 +42,12 @@ use crate::{
         chainspec_loader::ChainspecInfo,
         contract_runtime::{EraValidatorsRequest, ValidatorWeightsByEraIdRequest},
         fetcher::FetchResult,
-        linear_chain::FinalitySignature,
     },
     crypto::hash::Digest,
     rpcs::chain::BlockIdentifier,
     types::{
         Block as LinearBlock, Block, BlockHash, BlockHeader, Deploy, DeployHash, DeployHeader,
-        DeployMetadata, FinalizedBlock, Item, StatusFeed, Timestamp,
+        DeployMetadata, FinalitySignature, FinalizedBlock, Item, StatusFeed, Timestamp,
     },
     utils::DisplayIter,
     Chainspec,
